+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added extensible login actions to allow custom actions after a successful login
 * Added `bcmath` and `gmp` to suggested php extensions in `composer.json` (GitHub Issue #5)
 * Added configuration to automatically redirect users to the identity provider for login
+* Added ID-Token condition for OpenID Connect based providers (GitHub Issue #34)
 
 **Changed**